+++ src/react/atlascode/pullrequest/pullRequestDetailsController.ts
@@ -1,4 +1,4 @@
-import { MinimalIssue } from '@atlassianlabs/jira-pi-common-models';
+import { MinimalIssue } from '@atlassian-pi/jira-pi-common-models';
 import React, { useCallback, useMemo, useReducer } from 'react';
 import { defaultActionGuard, defaultStateGuard, ReducerAction } from 'src/ipc/messaging';
 import { v4 } from 'uuid';